requirements.txt
Flask
Werkzeug
requests
python-dotenv
Pillow
google-genai
